requirements.txt
python-docx
pyqt6